ast2600: Add IMA/EVM config variables to local.conf
IMA and EVM support requires Linux kernel support that can be enabled with
the DISTRO_FEATURE 'ima'. Add this to local.conf.

Beyond that, IMA and EVM require a couple of variables that the user should
set for her own needs, such as keys and certs used by IMA.

Have executables in the filesystem signed by inheriting the ima-evm-rootfs
IMAGE_CLASS.

Add tools getfattr and setfattr for getting and setting of xattrs. These
tools are useful to have in a package when using IMA and EVM signatures.

Change-Id: I7a13074e91a2928b1e5768e697c3f2c8430020d4
Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>