phosphor: Activate IMA and EVM in init script
Load IMA and EVM keys and activate EVM and load the IMA policy. When an
IMA appraisal policy is loaded, apply the IMA file signatures found
in the file 'ima_file_signatures' to files in the initrd. Also apply
the signatures to files that were copied from the initrd to
run/initramfs and that may be invoked upon machine shutdown.

Change-Id: I3cba35f6b7a8b5b3306a5d19daaa65fa185c7632
Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>