p10bmc: Add variables for IMA & EVM support
Add the variables needed for IMA support by requiring phosphor-ima.inc.

IMA and EVM support requires Linux kernel support that can be enabled with
the DISTRO_FEATURE:appen = ' integrity ima'. A user is expected to add
this to conf/local.conf.

Beyond that, IMA and EVM require a couple of variables that the user should
set for her own needs, such as keys and certs used by IMA. Those variables
should be set in conf/local.conf:
 - IMA_EVM_PRIVKEY: key used for signing files
 - IMA_EVM_ROOT_CA: the root CA used to verify signature verification keys
 - IMA_EVM_X509   : signature verification key

Have executables in the filesystem signed by activating the ima-evm-rootfs
by appending this to IMAGE_CLASS.

Add tools getfattr and setfattr for getting and setting of xattrs. These
tools are useful to have in a package when using IMA and EVM signatures.

Change-Id: I4ae648d2826b59f38f56b093164e40fb058b9092
Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>