mmc-init: Activate IMA and EVM in init script
Load IMA and EVM keys and activate EVM and load the IMA policy. When an
IMA appraisal policy is loaded, prepare for running signed busybox and
using signed libraries once the policy has been loaded. The signed
libraries are needed during the final switch_root launching 'init'.

Avoid activation of IMA and EVM if not running on QEMU and the secure
boot is disabled (jumper on motherboard) to allow installation of
(unsigned) tools for debugging.

Change-Id: I9c950f0bd1686b1b5ca94d2e16ee4032e457f33c
Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>