mmc-init: Activate IMA and EVM in init script as early as possible
Load IMA and EVM keys and activate EVM and load the IMA policy as soon
as the later root filesystem has been mounted. When an IMA policy is
loaded, have the tools (mount, switch_root, etc.) from the later root
filesystem in the PATH before the initrd ones because those ones may be
signed and work with an appraisal policy.

Change-Id: I9c950f0bd1686b1b5ca94d2e16ee4032e457f33c
Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>