mmc-init: Activate IMA and EVM in init script
Load IMA and EVM keys, activate EVM, and load the IMA policy. Before an
IMA appraisal policy is loaded, prepare the initramfs for running signed
executables and libraries by applying file signatures read from text
files produced during the image build.

Avoid activation of IMA and EVM if not running on QEMU and the secure
boot is disabled (jumper on motherboard) to allow installation of
(unsigned) tools for debugging. To allow disablement of IMA and EVM
without having to physically interact with the machine, provide
boot command line support for 'no-ima'.

Change-Id: I9c950f0bd1686b1b5ca94d2e16ee4032e457f33c
Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>