phosphor: Add features/ima/ima.scc
meta-security switched to using features/ima/ima.scc, so we have to
follow. Add ima.cfg and ima.scc to linux/features under under the
meta-phosphor/recipes-kernel directory so that it can be commonly used.

Add support for XATTRS to tmpfs, squashfs, and ext4 used by OpenBMC.
tmpfs needs xattr support so that files in the initrd can be signed.

Change-Id: I94532bdfa3818d164ace6e4de35f6238edc48cc7
Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>