phosphor: aspeed: Use rootfstype=tmpfs for boot when IMA enabled
When IMA is enabled, use the Linux boot parameter rootfstype=tmpfs to
have the initrd run in tmpfs, which allows file signatures to be applied
if tmpfs supports xattrs.

To use the Linux boot parameter for ast2600 adjust the u-boot default
configuration for this board with a config change. The config change has
to be applied to u-boot and also u-boot-fw-utils since the latter builds
the default configuration parameters into its fw_printenv tool and uses
this tool to overwrite the configuration on a running machine.

Change-Id: I55d094242176a44b85113ac645de7d5576ac645b
Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>