phosphor: Enable signing of files while in initrd
Enable the signing of files in the initrd by having meta-security's
ima_evm_sign_rootfs script write filenames and their IMA & EVM
signatures into 'ima_file_signatures' and 'evm_file_signatures'
respectively. Both files will be available in the initrd and enable
to sign executables and libraries when the initrd is run in tmpfs.
For this tmpfs needs to support xattrs. Having the files in the
initrd sign allows to enable an IMA signature-enforcing appraisal
policy already in the initrd.

Change-Id: Icd214ea08c37e52d54f645ec73d923743d5606fb
Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>