TEMP: meta-security: Write filenames and signatures into file
Note: This patch has to be upstreamed to meta-security.

Add two variables for filenames where the ima_evm_sign_rootfs script can
write the names of files and their IMA or EVM signatures into.

Having the filenames along with their signatures is useful for signing
files in the initrd when the initrd is running out of a tmpfs
filesystem that has support for xattrs. This allows to enable an IMA
appraise policy already in the initrd.

Change-Id: If5c5cc1b8622dc5ce27a065e921e19f57f411ca8
Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>